Add support for object classes and composite objects layers

+++ include/bntmx_map_object.h
@@ -14,7 +14,7 @@ namespace bntmx
 struct map_object
 {
     bn::fixed_point position;
-    int id;
+    uint16_t id;
 };
 
 }
